requirements.txt
colorama==0.4.6
docx2pdf==0.1.8
lxml==5.2.2
pillow==10.3.0
python-docx==1.1.2
pywin32==306
tqdm==4.66.4
typing_extensions==4.12.2
